Revert "Remove requirement of disabling  of NestJS"
This reverts commit 8fbc8121c0ddfdfc6c98457f477e47d0c6d1884b.

Apparently, converting the parsed webhook JSON back to JSON produces differences in the raw webhook body.
For instance:

* a `/` is escaped by Shopify, but not escaped by `JSON.stringify`.
* monetary values are always suffixed with `.0`, but not when you call `JSON.stringify`

this was causing the HMAC signing to fail. In other words, you *need* to disable the bodyParser
logic of NestJS in order to access the raw webhook body to calculate the HMAC.